Add read capability to the SocketClient (new name of SocketWriter)
